Add virtual package resolution
Let's talk about this monster of a commit.

First thing we do is add a new architecture to the graph
It's really just a map from virtual? maps to real package names.
We tell bdot to ignore these, by the way.

We add a module for parsing virtual pkgs in a friendly way. We call
it bvirtpkgs. We read in an array of strings: all the even indexes
(starting with 0) are the virtual names, and the odd indexes are
the relevant packages they go to.

We read these in the pkgimport_master, and we do NOT save them to the
database. This means the pkgimport_master needs to be online for the
grapher to work at all. But this was already a hard-coded requirement.
Oh, and the virtual package map is transient. It's used only for package
resolution (going from names to pkg structs) and never again, so every
time pkgimport_master gets new virtual packages (or the same list again)
we just destroy the old list and re-add it to the graph. Not a big deal.

We did need to add a constant BGRAPH_VPKG_STR, to placehold as the
"arch" that's actually all the virtual package maps. And to send the
messages, we added the HEREVIRTUALPKGS message to the pkgimport chain.
There are also no tests which respect the new need for a virtual package
list when operating dxpb. This will need fixing in the future.