Trigger error in useblog
This is part of how dxpb uses this, since we don't yet bind workers
and which packages they are building, not in a way that's available to
the frontend on LOGHERE. Not good thing, but to be fixed later.

$ make useblog
$ ./useblog # No error here.
$ capnp decode --packed log.capnp LogEntry \
	< /tmp/dxpb-blog-test-*/capnp.packed > message.txt
*** ERROR DECODING PREVIOUS MESSAGE ***
The following error occurred while decoding the message above.
This probably means the input data is invalid/corrupted.
Exception description: expected boundsCheck(segment, ptr, \
	ref->structRef.wordSize()); Message contained out-of-bounds \
	struct pointer.
Code location: capnp/layout.c++:2159
*** END ERROR ***
$ echo "Now the log is uselss to a user."