Allow a package to be tried multiple times
The 'bad' field on a package, which is not persistent past reboots or
package reads, is now an integer field instead of a bit. It provides
a counter for us, currently stopping at 3, to see how many times we
gave up on building a particular package. The log message about
a package being marked bad will only be generated when the package has
been marked bad for the third time.