Check that ARCH_HOST deps are present
Also in this commit: To prevent from linking against all of tbe grapher
in the dxpb-frontend, we split out worker matching (the package dating
service where we set up eligible young packages with suitable builders)
into a new, separate, module.

Basically all we do is make sure the ARCH_HOST dependencies are actually
in the repo, when we know what the host architecture of the builder is
going to be.