Problem: No documentation
Solution: Add a skeleton of requirements, allowing us to know when dxpb
is ready for any release.

The goal is the AMVP right now. That will be enough to replace buildbot,
and only replace buildbot. It replaces no problems except the complexity
of buildbot with a tailored solution, and allows us to do nothing except
upgrade to better and more featureful dxpb systems.