Problem: Existing IRC Implementation is not Maintainable
Solution: Rewrite the IRC implementation

The existing IRC implementation has a few problems.  The first is that
the package does not actually own its configuration values.  This is
resovled by removing the defaults from the main package, and then
re-homing those defaults to the irc package.

Another problem with the existing implementation is rooted in the use
of a global variable which stores the IRC connection.  This is
resolved by using a proxy type which encapsulates the IRC connection
and provides the convenience functions the rest of the application
will use to interact with IRC.

Finally the irc package does not follow the Go convention of returning
errors to the calling package and allowing the caller to decide if
initialization of a subpkg is fatal.  This has been rectified by
returning all reasonable errors to the calling package, and in the
singular case where no error should return returning an uncatchable
error.  I chose not to make structural changes to main, but it may
be prudent to wrap irc.Connect() in a panic() to handle this error.