Problem: Need to run more than xbps-src on the builder
Solution: add a shell script based on the current buildbot config.

This adds assets/builder/xbps-bulk-builder.sh, which is just a
standalone wrapper around xbps-bulk which can pull its own
void-packages and bootstrap a build environment, if one is not
alrady preset. It can pass through flags to xbps-bulk, -N to
disable remote dependency resolution, -t to create a temporary
overlay of the masterdir, and -R to add a repo to the top of the
list for xbps-checkvers.

-d, -m, and -h specify paths to dist (void-packages), master,
and host directories. -b specifies a directory for each build,
or defaults to '$PWD/builddir' and cleans the directory before
running.

-c copies a config file to xbps-src's etc/conf in void-packages.

Cross targets need their respective repodata cache. A simple way
to get this is 'XBPS_TARGET_ARCH=arch xbps-install -S' for
testing. Also, mixing glibc/musl hosts and bootstraps is finnickey
and I have not been able to document a stable procedure for this.