Problem: Can't trigger build
Solution: Add broadcasting and trigger on stdin. It works for now
and we just need a proof of concept for right now. Later we can
hook up the channel to a github webhook or IRC. Also later, we
will transform jobs to only trigger specific workers, but that logic
need not be too complicated.

The (future) graph need not be aware that its jobs are being handled. It
just has to keep track of packages. Life is better that way.